+++ secureli/actions/build.py
@@ -1,8 +1,8 @@
 from secureli.modules.shared.abstractions.echo import EchoAbstraction, Color
 from secureli.modules.observability.observability_services.logging import (
     LoggingService,
-    LogAction,
 )
+from secureli.modules.shared.models.logging import LogAction
 
 
 class BuildAction:
