+++ secureli/actions/update.py
@@ -3,12 +3,12 @@ from pathlib import Path
 from secureli.modules.shared.abstractions.echo import EchoAbstraction
 from secureli.modules.observability.observability_services.logging import (
     LoggingService,
-    LogAction,
 )
 from secureli.modules.core.core_services.updater import UpdaterService
 from secureli.actions.action import Action, ActionDependencies
 
 from rich.progress import Progress
+from secureli.modules.shared.models.logging import LogAction
 
 
 class UpdateAction(Action):
