+++ secureli/modules/language_analyzer/language_analyzer_services/language_analyzer.py
@@ -1,31 +1,12 @@
 from collections import defaultdict
 from pathlib import Path
 
-import pydantic
-
 from secureli.modules.shared.abstractions.lexer_guesser import LexerGuesser
+from secureli.modules.shared.models.language import AnalyzeResult, SkippedFile
 from secureli.repositories.repo_files import RepoFilesRepository
-from secureli.modules.language_analyzer.language_analyzer_services.language_support import (
+from secureli.modules.shared.consts.language import (
     supported_languages,
 )
-
-
-class SkippedFile(pydantic.BaseModel):
-    """
-    A file skipped by the analysis phase.
-    """
-
-    file_path: Path
-    error_message: str
-
-
-class AnalyzeResult(pydantic.BaseModel):
-    """
-    The result of the analysis phase.
-    """
-
-    language_proportions: dict[str, float]
-    skipped_files: list[SkippedFile]
 
 
 class LanguageAnalyzerService:
