+++ secureli/modules/shared/consts/language.py
@@ -1,0 +1,11 @@
+supported_languages = [
+    "C#",
+    "Python",
+    "Java",
+    "Terraform",
+    "TypeScript",
+    "JavaScript",
+    "Go",
+    "Swift",
+    "Kotlin",
+]
