+++ tests/actions/test_initializer_action.py
@@ -5,13 +5,8 @@ import pytest
 
 from secureli.actions.action import ActionDependencies
 from secureli.actions.initializer import InitializerAction
-from secureli.repositories.secureli_config import SecureliConfig
-from secureli.repositories.settings import SecureliFile, TelemetrySettings
-from secureli.modules.language_analyzer.language_analyzer_services.language_config import (
-    LanguageNotSupportedError,
-)
-from secureli.modules.observability.observability_services.logging import LogAction
-from secureli.settings import Settings
+from secureli.modules.shared.models.language import LanguageNotSupportedError
+from secureli.modules.shared.models.logging import LogAction
 
 test_folder_path = Path("does-not-matter")
