+++ tests/application/test_main.py
@@ -4,12 +4,12 @@ from typer.testing import CliRunner
 
 import pytest
 from pytest_mock import MockerFixture
-from secureli.actions.action import VerifyOutcome, VerifyResult
 
 import secureli.container
 import secureli.main
+from secureli.modules.shared.models.install import VerifyOutcome, VerifyResult
 from secureli.modules.shared.models.publish_results import PublishResultsOption
-from secureli.modules.core.core_services.scanner import ScanMode
+from secureli.modules.shared.models.scan import ScanMode
 from secureli.modules.shared.utilities.secureli_meta import secureli_version
